put oracles page back
